WHF-227: Add support for smart groups
We used GroupContact API to fetch the groups that the current logged
user belongs to, but this will not work with smart groups. This commit
adds support for smart groups by querying the table
`civicrm_group_contact_cache`, I chose the raw query because there is no
official API for that.